Added namespacing for multi-instance runs.
In order to make multiple instances able to run side-by-side, they must
not not overlap topics they are publishing to. With this commit, each
instance of last_letter is namespaced as an anonymous node with a
randomly (?) generated name, with the <group> tag in launcher.launch.

NB: The clock node was namespaced by simply moving it out of the global
namespace (being published to /clock) into the local, by removing the
leading "/". This seems to be fine.