Moved clock to root namespace.
clock node now checks for a flag existing an indicating clock node. If
one is not found, or if it is set to false, it begins to publish on
/clock, and sets the flag to true.

TODO: Find a way to check if there is a /clock node with any publishers
without using the flag to support clocks launched through other ROS
software besides last_letter.